Generate cmp_test_results in multiple directories (#6)
Generate cmp_test_results in any directory, and to simplify actions with
macros, move go test + replacement logic into a separate shell script.

The testdata files are only regenerated if `*_test.go` files in the
directory change (or the cmptest go files change).
To force a rerun, add a `force-update-testdata` target,
which is an alias for `make -B update-testdata`.