Verify that `fakeTB` implements all testing.TB methods explicitly
`fakeTB` embeds `testing.TB` since the interface contains a `private()`
method which we cannot implement. Embedding works around this, but also
means all methods on the interface are available on `fakeTB` even if we
did not expicitly implement them.

Since the embedded `testing.TB` value is left `nil`, accesses to this
method will panic. To avoid new methods in the interface going
unimplemented and causing panics, the added test ensures that `fakeTB`
explicitly implements every method of `testing.TB`.

Reflection does not differentiate methods on `fakeTB` that are
implemented on the type vs those from the embedded interface, so use AST
parsing to find explicitly implemented methods.