exp/retryt: Retry a `func(testing.TB)` till it passes (#12)
Add an experimental package that uses faket to implement a helper to
retry test assertions till they succeed.

This is intended for assertions against asynchronous work which would be
flaky without a wait. By using retryt, tests can avoid custom code that
waits for a condition independent of the test assertions.