Implement `t.Helper()` to skip helper frames
Store the caller PC when `Helper` is called, and use this
to generate a list of functions to skip.

The implementation aligns with how `testing.TB` skips helpers based
on function rather than PC directly, likely to avoid issues with
inlining, which can result in shared PC values for different functions.

Expose the list of helper functions via `TestResult.Helpers()` as
it's required internally, and may be useful to verify that a helper
correctly marks itself as a helper.