Fix handling of `t.Skip` inside of `t.Cleanup` funcs (#16)
Skips uses `runtime.Goexit()` to exit the goroutine but is still
expected to run cleanup functions. Use `defer` as exiting will
run deferred calls, and use it to continue running cleanups.

Update the test to mix nesting and skips.

Note: The `defer nested cleanup` log has an incorrect caller
`panic.go, but this aligns with stdlib.