Refactor Logs APIs and implementation (#20)
`TestResult.Logs()` now returns a list of structured entries for logs.

Most callers will likely rely on `Logs.String()` to get a concatenated
string of all logs, in a format similar to `go test`. However, tests can
also get a list of messages, or work with the `Log` entry directly.

The log entry includes the message, caller information, and the
`testing.TB` function that was used to log. In the future, we can
include caller PCs if there's use-cases, but PCs are not exposed
due to the complexity of merging caller PCs across the log caller,
and possibly a cleanup caller.