Fix panics setting TestResult asynchronously (#36)
Panic handling is deferred before the defer that marks tests complete.
This could lead to a test being marked complete, returning a result but
then mutating that result after returning.

This happened infrequently but could lead to a flaky result where a
panicked test isn't marked as `Failed()`. Added test runs many
iterations so we are more likely to trigger a failure on flakes.